requirements.txt
Django>=4.2
gunicorn
psycopg2-binary
celery
redis
django-environ
drf-yasg
whitenoise
djangorestframework
djangorestframework-simplejwt
django-redis
django-celery-beat
django-cors-headers
pytest
pytest-django
pytest-cov
factory-boy
